Updated script.js, style.css and index.html
Scipt bug that caused each new search to just add to the list of forecast has now been fixed. It will only show the forecast for the most recent searched city. The seach history list is getting to long now though. Updated html and css to fix styling so it's not all displayed in one long list